% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{BotonEstado}
\alias{BotonEstado}
\title{Boton interruptor (toggle switch)}
\usage{
BotonEstado(...)
}
\arguments{
  \item{...}{Argumentos pasados a `shinyWidgets::materialSwitch`.}
}
\value{Componente Shiny.}
\description{
Boton interruptor (toggle switch)
}
\examples{
\dontrun{
BotonEstado("toggle_filtro", "Activar filtro")
}
}
