% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{CajaIco}
\alias{CajaIco}
\title{Caja informativa con icono y color personalizable}
\usage{
CajaIco(...)
}
\arguments{
  \item{texto}{Texto principal a mostrar.}
  \item{icono}{Nombre del icono FontAwesome.}
  \item{col_fondo}{Color de fondo de la caja.}
  \item{alto}{Alto de la caja en pixeles.}
  \item{col_letra}{Color del texto.}
  \item{col_icono}{Color del icono.}
}
\value{Objeto HTML de Shiny.}
\description{
Caja informativa con icono y color personalizable
}
\examples{
\dontrun{
CajaIco("Ingresos totales", "chart-line")
}
}
