% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{chr_kpi}
\alias{chr_kpi}
\title{Indicadores textuales de desempeno KPI}
\usage{
chr_kpi(...)
}
\arguments{
  \item{x}{Vector numerico de cumplimiento.}
}
\value{Vector de caracteres con flechas indicadoras HTML.}
\description{
Indicadores textuales de desempeno KPI
}
\examples{
chr_kpi(c(0.7, 1.0, 1.2))
}
