% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{col_num}
\alias{col_num}
\title{Paleta numerica estandar del paquete}
\usage{
col_num(...)
}
\arguments{
  \item{x}{Vector numerico para mapear a colores.}
}
\value{Vector de colores hexadecimales (escala verde-azul).}
\description{
Paleta numerica estandar del paquete
}
\examples{
col_num(1:5)
}
