% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{DefinirFormato}
\alias{DefinirFormato}
\title{Definir formato preconfigurado}
\usage{
DefinirFormato(...)
}
\arguments{
  \item{formato}{Nombre del formato.}
  \item{fn}{Funcion de formato personalizada (opcional).}
  \item{...}{Argumentos adicionales pasados al generador de `scales`.}
}
\value{Funcion de formato.}
\description{
Definir formato preconfigurado
}
