% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{FormatoHOT}
\alias{FormatoHOT}
\title{Convertir un formato registrado a sintaxis Handsontable}
\usage{
FormatoHOT(...)
}
\arguments{
  \item{formato}{Nombre del formato registrado.}
}
\value{Cadena con el patron de formato para Handsontable.}
\description{
Convertir un formato registrado a sintaxis Handsontable
}
