% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{FormatoJS}
\alias{FormatoJS}
\title{Convertir un formato registrado a representacion JavaScript}
\usage{
FormatoJS(...)
}
\arguments{
  \item{formato}{Nombre del formato registrado.}
}
\value{Cadena con funcion JS como string.}
\description{
Convertir un formato registrado a representacion JavaScript
}
