% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{gt_minimal_style}
\alias{gt_minimal_style}
\title{Aplicar estilo minimalista a tabla gt}
\usage{
gt_minimal_style(...)
}
\arguments{
  \item{gt_table}{Objeto `gt_tbl`.}
}
\value{Objeto `gt_tbl` con estilo aplicado.}
\description{
Aplicar estilo minimalista a tabla gt
}
