% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{gt_var_style}
\alias{gt_var_style}
\title{Formato variacional (+/-) en columnas gt}
\usage{
gt_var_style(...)
}
\arguments{
  \item{gt_table}{Objeto `gt_tbl`.}
  \item{...}{Columnas seleccionadas.}
}
\value{Objeto `gt_tbl` con formato variacional.}
\description{
Formato variacional (+/-) en columnas gt
}
