% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{InputMes}
\alias{InputMes}
\title{Selector de mes (wrapper de `InputFecha`)}
\usage{
InputMes(...)
}
\value{`shiny.tag` del widget `airDatepickerInput` configurado para meses.}
\description{
Selector de mes (wrapper de `InputFecha`)
}
