% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{ObtenerFormato}
\alias{ObtenerFormato}
\title{Obtener un formato registrado por nombre}
\usage{
ObtenerFormato(...)
}
\arguments{
  \item{nombre}{Nombre del formato.}
}
\value{Funcion de formato.}
\description{
Obtener un formato registrado por nombre
}
