% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{pick_opt}
\alias{pick_opt}
\title{Construir opciones con genero para pickerInput}
\usage{
pick_opt(...)
}
\arguments{
  \item{cho}{Vector de opciones.}
  \item{fem}{Logico. `TRUE` para opciones en femenino.}
}
\value{Lista de opciones para `shinyWidgets::pickerOptions`.}
\description{
Construir opciones con genero para pickerInput
}
